// src/suport/enums/bookingEngine.enum.ts
export enum BookingRulesCheckboxes {
    LIMIT_CONCURRENT_BOOKINGS = "Limit Concurrent Bookings",
    LIMIT_CONCURRENT_HOURS = "Limit Concurrent Hours",
    LIMIT_DAILY_PLAY = "Limit Daily Play",
    LIMIT_MONTHLY_PLAY = "Limit Monthly Play"
}

export enum BookingRulesInputs {
    MAX_CONCURRENT_BOOKINGS = "Max Concurrent Bookings",
    MAX_CONCURRENTLY_BOOKED_HOURS = "Max Concurrently Booked hours",
    MAX_HOURS_PER_DAY = "Max Concurrently Booked hours",
    MAX_HOURS_PER_MONTH = "Max Hours per Month"
}